+++ splash-screen/src/overlay/traits.rs
@@ -6,10 +6,10 @@ pub trait Overlay {
     fn z_index(&self) -> u32;
 
     /// Draw the `Overlay` into the given `Frame` for the specified timestamp.
-    fn draw(&self, frame: &mut Frame, timestamp_ms: u64);
+    fn draw(&self, frame: &mut Frame, timestamp_ms: u128);
 
     /// Time in milliseconds until this overlay wants the next `Frame`.
     /// - Return `None` if this overlay does not have a timed next frame (static image).
     /// - Return `Some(remaining_ms)` where `remaining_ms` is >= 0.
-    fn time_to_next_frame_ms(&self, timestamp_ms: u64) -> Option<u64>;
+    fn time_to_next_frame_ms(&self, timestamp_ms: u128) -> Option<u128>;
 }
